[TLC-254] fix issue where vals stripped fields are cfg'd more than once
If a field like dc.title.subtitle is configured multiple times in one
form with different rules for different type binding, when the validation
encounters the input config that is *not* bound for the doc type, it will
strip all metadata values even though the same field is allowed for this
type elsewhere in the form. This commit should fix this by constructing
a lookup map first, though it does not have a way to rigorously enforce
specific rule conflicts like repeatable=true and repeatable=false for
the same type, and so on.