Explicit property must be removed from test configuration too.
I changed over from field injection to constructor injection while
diagnosing the failure caused by this problem, and decided to keep
it.  It's a little more verbose, but safer and clearer in the doco.
Really, the *required* properties should also be constructor
parameters.

Also cleaned up some trailing whitespace in the configuration documents.