Fix OpenSearch NullPointerException for unknown valid UUIDs in scope
Fixes a NullPointerException when the "scope" parameter provided to
the OpenSearch endpoint is a valid UUID, but is not a UUID associated
with a Community or Collection.

Instead of throwing a NullPointerException, this change modifies the
code to return a null scope (resulting in an "unscoped" OpenSearch
request), which is the same behavior that occurs when the UUID
is invalid, or otherwise not usable.